Add stub capture for testing
After the script is written, perhaps before the refactor, I'd like
to try writing some tests that call `build` repeateadly to test how
the script's output compares to that of the real `build` command.

We are working around the limitations of `vftool` by treating the
virtual filesystem as a registry (which it is), manually cleaning
its text representation with our script, and importing it via the
`vregtool`. Something is bound to go wrong somewhere.

It seems that I cannot add the empty `exe` file to git, so I'll
have to come up with some other way to test the app. Maybe the
`cmd` entry point is enough.